+++ inspectors/store/appmedia.go
@@ -25,8 +25,8 @@ import (
 	"net/url"
 
 	. "github.com/canonical/fetch-service/inspectors/common"
+	"github.com/canonical/fetch-service/inspectors/mimetypes"
 	"github.com/canonical/fetch-service/inspectors/store/config"
-	"github.com/canonical/fetch-service/metadata/opinions"
 )
 
 type StoreAppMediaInspector struct {
@@ -81,13 +81,14 @@ func (ins *StoreAppMediaInspector) inspectArtifactPNG(f ArtifactReader, a Respon
 	}
 
 	md := ArtifactMetadata{
+		Type:        mimetypes.StoreAppmediaPNG,
 		Name:        "Image file",
 		Description: "Store media file in PNG format",
 	}
 
 	a.SetArtifactMetadata(md)
 
-	if a.InspectorRequestOpinion(ins) == opinions.Pending {
+	if a.InspectorRequestOpinionPending(ins) {
 		a.SetResponseApproved(ins, "store media file in PNG format").Annotate(notes)
 	} else {
 		a.SetResponseUnknown(ins, "unknown PNG image").Annotate(notes)