+++ service/handlers_test.go
@@ -53,7 +53,7 @@ func (t *serviceSuite) TestReuseInspectionResult(c *C) {
 	}
 
 	// Start session
-	s := session.New(opt.Spool, 0, true, nil, config.SessionInspectorsConfig{})
+	s := session.New(opt.Spool, 0, true, nil, config.OverrideInspectorsConfig{})
 	defer s.Discard()
 
 	// Create a fake artifact
@@ -164,7 +164,7 @@ func (t *serviceSuite) TestCompleteInspectionAfterSessionFinished(c *C) {
 	}
 
 	// Start session
-	s := session.New(opt.Spool, 0, true, nil, config.SessionInspectorsConfig{})
+	s := session.New(opt.Spool, 0, true, nil, config.OverrideInspectorsConfig{})
 	defer s.Discard()
 
 	// Create a fake artifact